+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Version 2.2.4
 
+* [165](https://github.com/mlebreuil/netbox-contract/issues/160) Add more fields to the invoice and contract bulk edit forms.
 * [165](https://github.com/mlebreuil/netbox-contract/issues/165) Fix Invoice and invoiceline creation through api. 
 
 ### Version 2.2.3